+++ src/Caffeine/Classes/Views/PreferencesView.swift
@@ -13,6 +13,7 @@ struct PreferencesView: View {
     @AppStorage(PreferenceKeys.activateAtLaunch) private var activateAtLaunch = false
     @AppStorage(PreferenceKeys.suppressLaunchMessage) private var suppressLaunchMessage = false
     @AppStorage(PreferenceKeys.deactivateOnManualSleep) private var deactivateOnManualSleep = false
+    @AppStorage(PreferenceKeys.automaticLidControl) private var automaticLidControl = true
     @AppStorage(PreferenceKeys.keepAppsActive) private var keepAppsActive = false
 
     var body: some View {
@@ -97,10 +98,14 @@ struct PreferencesView: View {
                     .padding(.leading, 20)
             }
 
-            Divider().padding(.vertical, 12)
-            AutomaticLidControls(manager: self.viewModel.automaticLid, isActive: self.viewModel.isActive) {
-                self.viewModel.toggleActive()
-            }
+            Divider().padding(.vertical, 8)
+            AutomaticLidControls(manager: self.viewModel.automaticLid, isEnabled: Binding(
+                get: { self.automaticLidControl },
+                set: { enabled in
+                    self.automaticLidControl = enabled
+                    self.viewModel.updateAutomaticLidControl(enabled: enabled)
+                }
+            ))
 
             Spacer()
                 .frame(height: 30)